Got pathplanning turning over
There's a substantial bug / usabiliity issue in the library we're building on top of,
to the point where the demo code on the package website is wrong.  It's a pretty simple
problem with object equality using addrs vs the intended behavior of using node location.

I could use a different library, code around the issue (what I've
done) or try to get a patch merged in upstream.  It hasn't been
updated in years, so I think I'll use this as a teaching beat.